A first version of the new download command.
This is intended to override rebar to avoid issues around the broken behaviour
with conflicting versions in dependencies. For more details, see the rebar
issue here:
 * https://github.com/rebar/rebar/issues/170

Incidentally, this is no indictment of the rebar team -- they've got a huge
maintenance burden hanging over them, and are going great work! We just need a
fix for this badly enough that we'll write a tool for it :-)

This commit is in support of issue #135.